requirements.txt
py-cord[voice]==2.6.1
openai>=1.40.0
python-dotenv>=1.0.0
RealtimeSTT>=0.3.0
RealtimeTTS>=0.4.0
PyNaCl>=1.5.0
Pillow>=10.0.0
aiohttp>=3.9.0
numpy>=1.24.0
